backend/requirements-dev.txt
-r requirements.txt
pytest==8.0.2
pytest-django==4.8.0

